SpotInfo DB 접근 시 권한 체크, UserPrincipal 코드 일부 수정

+++ TravelNode/src/main/java/com/example/travelnode/controller/SpotController.java
@@ -1,12 +1,16 @@
 package com.example.travelnode.controller;
 
 import com.example.travelnode.dto.SpotInfoDto;
+import com.example.travelnode.entity.RoleType;
 import com.example.travelnode.entity.SpotInfo;
+import com.example.travelnode.oauth2.entity.UserPrincipal;
 import com.example.travelnode.service.SpotService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Objects;
 
 @RestController
 @RequestMapping("/api/v1/spot")
@@ -22,20 +26,46 @@ public class SpotController {
     }
 
     @PostMapping("/register")
-    public SpotInfo registerSpot(SpotInfoDto spotInfoDto) {
+    public SpotInfo registerSpot(@AuthenticationPrincipal UserPrincipal userPrincipal,
+                                 @RequestBody SpotInfoDto spotInfoDto) throws Exception {
+
+        if(userPrincipal == null) {
+            throw new Exception("No User Information");
+        }
+
+        if(Objects.equals(userPrincipal.getRoleType().getCode(), RoleType.USER.getCode())) {
+            throw new Exception("Do not have authority");
+        }
 
         return spotService.registerSpotInfo(spotInfoDto);
     }
 
-    @PutMapping("/change-name")
-    public SpotInfo updateSpotName(String prevName, String spotName) {
+    @PatchMapping("/change-name")
+    public SpotInfo updateSpotName(@AuthenticationPrincipal UserPrincipal userPrincipal,
+                                   @RequestParam String prevName, @RequestParam String spotName) throws Exception {
+
+        if(userPrincipal == null) {
+            throw new Exception("No User Information");
+        }
+
+        if(Objects.equals(userPrincipal.getRoleType().getCode(), RoleType.USER.getCode())) {
+            throw new Exception("Do not have authority");
+        }
 
         return spotService.updateSpotName(prevName, spotName);
     }
 
     @DeleteMapping("/delete/{spotName}")
-    public void deleteSpot(@PathVariable(value = "spotName") String spotName) {
+    public void deleteSpot(@PathVariable(value = "spotName") String spotName,
+                           @AuthenticationPrincipal UserPrincipal userPrincipal) throws Exception {
 
+        if(userPrincipal == null) {
+            throw new Exception("No User Information");
+        }
+
+        if(Objects.equals(userPrincipal.getRoleType().getCode(), RoleType.USER.getCode())) {
+            throw new Exception("Do not have authority");
+        }
         spotService.deleteSpotInfo(spotName);
     }
 }
